getHumidite et toutes les données ok

+++ app/src/main/java/com/example/pageacceuil/Data.java
@@ -3,8 +3,8 @@ package com.example.pageacceuil;
 
 public class Data {
 
-    private float humidite;
-    private float temperature;
+    private double humidite;
+    private double temperature;
     private float lux;
     private float O2;
     private float CO2;
@@ -13,7 +13,7 @@ public class Data {
     public Data() {
     }
 
-    public float getHumidite() {
+    public double getHumidite() {
         return humidite;
     }
 
@@ -21,7 +21,7 @@ public class Data {
         this.humidite = humidite;
     }
 
-    public float getTemperature() {
+    public double getTemperature() {
         return temperature;
     }
